Moved string splitting functions; added string character removal function
